Consolidating sendResponse into one method in the DefaultHandler class
